delete resultView class and insert QuizSystem class, add animation, fix CLI display
